fix: increase timeout for aggregator API (#104)
All handler executions for inclusion store insert are timing out at 10s. The lambda invokes the aggregator service UCAN service, but this is also set to timeout at the default 10s.

<img width="785" alt="Screenshot 2024-07-17 at 11 42 46" src="https://github.com/user-attachments/assets/008e9d41-89f5-4261-ae6e-60227363e801">

This PR changes the timeout for the insert handler lambda and the aggregator UCAN service to the max (30s) to see if it resolves the issue.